'ARTIFICIAL COMMIT FOR NEW FILES'

+++ dataset/GitHub_Java/androidannotations.androidannotations/AndroidAnnotations/androidannotations-core/androidannotations-api/src/main/java/org/androidannotations/annotations/Background.java
@@ -1,6 +1,5 @@
 /**
- * Copyright (C) 2010-2016 eBusiness Information, Excilys Group
- * Copyright (C) 2016-2020 the AndroidAnnotations project
+ * Copyright (C) 2010-2015 eBusiness Information, Excilys Group
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -192,7 +191,7 @@ public @interface Background {
 	 *
 	 * @return the delay of the execution
 	 */
-	long delay() default 0;
+	int delay() default 0;
 
 	/**
 	 * Serial execution group.
